+++ phase4/systemd.sh
@@ -131,15 +131,43 @@ if [[ "$MULTILIB" == "true" ]]; then
 	cp -av DESTDIR/usr/lib32/libudev.so* /usr/lib32/
 	cp -v  DESTDIR/usr/lib32/pkgconfig/* /usr/lib32/pkgconfig/
 	rm -rf DESTDIR
-	
+fi
+if [[ "$MULTILIB" == "true" ]] && [[ "$MULTILIB_mx32" == "true" ]]; then	
 	#x32bit
 	rm -rf *
+
+	# Make a cross compile file or add syscall.x32=y to kernel commands
+cat > x32-cross.ini << "EOF"
+[binaries]
+c = 'gcc'
+cpp = 'g++'
+ar = 'ar'
+strip = 'strip'
+pkgconfig = 'pkg-config'
+
+[built-in options]
+c_args = ['-mx32', '-march=x86-64']
+c_link_args = ['-mx32']
+cpp_args = ['-mx32', '-march=x86-64']
+cpp_link_args = ['-mx32']
+
+[properties]
+# This tells Meson the compiled binaries cannot be run natively
+needs_exe_wrapper = true
+
+[host_machine]
+system = 'linux'
+cpu_family = 'x86_64'
+cpu = 'x86_64'
+endian = 'little'
+EOF
 	
 	PKG_CONFIG_PATH="/usr/libx32/pkgconfig" \
 	CC="gcc -mx32"                       \
 	CXX="g++ -mx32"                      \
 	LANG=en_US.UTF-8                     \
-	meson setup ..                       \
+	meson setup .. 						 \
+		  --cross-file x32-cross.ini     \
 		  --prefix=/usr                  \
 		  --libdir=/usr/libx32           \
 		  --buildtype=release            \
